合并SignerUtilsUtils -> AppUtils
